Correctly process pipelines
Pipelines are now updated based on the status of their builds. This process includes checking the current stage, enqueuing the next stage, making sure that failed builds mark the pipeline as failed, unless they are allowed to fail, and mark the pipeline as success if certain conditions are met.

This worker is enqueued every time a build is created/updated.