Remove Exq and replace it with Que
For what we used background processing at the moment, it was overkill to use Exq, specially because it depends on Redis. We want lower the barrier of entry for the project, and having to also maintain a Redis instance is too much.

Que uses Mnesia as a backend, wich comes by default with the BEAM, making perfect for our use cases.