Updating stupid fucking JS deps
For some stupid reason (because JS) node-sass v3.8.0 disappeared and would not allow the modules to compile. Distillery would not be able to find the correct version. It took way too long to fix. Also, for some reason, the fucking version of sass-brunch that is on npm is different than the one on GitHub, even though they carry the same fucking version number. If you added new code since the last version, and it has been more than a month, release a fucking new version already! (╯°□°）╯︵ ┻━┻