Improve query that fetches latest projects' statuses
The old way resulted in one query per shown project, which will lead to N+1 queries if the shown projects are to many. This fetches the statuses for the shown projects in a single query and outputs it into a map that uses the project_id as key. If no key is found, "unknown" is returned.